pyproject.toml
[project]
name = "travelagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.2",
    "google-generativeai>=0.8.5",
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.1",
]
